requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Faker
numpy
pandas
psycopg2